// src/lib/config/types.ts
export interface ResourceItem {
  id: number;
  name: string;
  description: string;
  category: string;
  cloudType: string;
  link: string;
  image: string | null;
  createdAt?: string;
}